+++ back/package.json
@@ -2,10 +2,10 @@
   "name": "soundhub",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "server.js",
   "type": "module",
   "scripts": {
-    "start": "nodemon src/app.js",
+    "start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -26,5 +26,9 @@
     "socket.io": "^4.2.0",
     "youtube-search-api": "^1.0.9",
     "zod": "3.21.4"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0",
+    "supertest": "^7.0.0"
   }
 }
